Add force and clean desktop file

+++ download-qupath-version.py
@@ -10,10 +10,16 @@ parser = argparse.ArgumentParser(
     prog="DownloadQuPath", description="Download a specified version of QuPath"
 )
 parser.add_argument(
+    "-f",
+    "--force",
+    help = "Overwrite existing downloads",
+    default = False,
+)
+parser.add_argument(
     "-v",
     "--version",
-    help="The version to download, eg v0.4.3, latest",
-    default="latest",
+    help = "The version to download, eg v0.4.3, latest",
+    default = "latest",
 )
 args = parser.parse_args()
 version = args.version
@@ -42,18 +48,26 @@ release = [release for release in releases if release["tag_name"] == version][0]
 linux_asset = [asset for asset in release["assets"] if "Linux" in asset["name"]]
 download_url = linux_asset[0]["browser_download_url"]
 
-tarfile = f"qupath-{version}.tar.xz"
-wget.download(download_url, tarfile)
-os.system(f"tar -xJf {tarfile}")
+if os.path.exists(version):
+    if not args.force:
+        print(f"Not overwriting existing download: {version}")
+    else:
+        print(f"Overwriting existing {version}")
+        tarfile = f"qupath-{version}.tar.xz"
+        wget.download(download_url, tarfile)
+        print("")
+        os.system(f"tar -xJf {tarfile}")
 
-os.system(f"rm -rf {version}")
-os.rename("QuPath", version)
+        os.system(f"rm -rf {version}")
+        os.rename("QuPath", version)
 
-os.remove(tarfile)
+        os.remove(tarfile)
+        
 
 if latest:
-    os.remove("latest")
-    os.symlink(f"{version}/bin/QuPath", "latest")
+    if os.path.exists("latest"):
+        os.remove("latest")
+    os.symlink(f"{version}", "latest")
     os.chmod("latest", 0o755)
 
     desktop = f"""
@@ -62,13 +76,13 @@ if latest:
     Version={version}
     Type=Application
     Terminal=false
-    Exec={os.path.realpath(f"latest")}
+    Exec={os.path.realpath(f"latest/bin/QuPath")}
     Name=QuPath
     Icon={os.path.realpath(f"{version}/lib/QuPath.png")}
     """
-    with open(
-        os.path.expanduser("~/.local/share/applications/qupath.desktop"), "w"
-    ) as f:
+    ## probably a better way to ensure there's no leading tabs, but...
+    desktop = "\n".join([x.strip() for x in desktop.split("\n")])
+    with open(os.path.expanduser("~/.local/share/applications/qupath.desktop"), "w") as f:
         f.write(desktop)
